backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
tenacity==9.0.0
python-dotenv==1.0.1
google-generativeai==0.8.3
google-search-results==2.4.2   # serpapi client
beautifulsoup4==4.12.3
lxml==5.3.0
torch>=2.6.0                   # compatible with Python 3.13
numpy>=1.26.0
orjson==3.10.7
cachetools==5.5.0